+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on a modified version of [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.6.3
+
+### News
+
+* Support for loading data into a model using `load()`.
+* Return a promise when preloading data using `preload()`.
+* Improved test coverage.
+
 ## 0.6.2
 
 ### News